Get rid of JaegerSpanContextTest
